+++ src/NCronJob/JobExecutionContext.cs
@@ -6,21 +6,19 @@ namespace NCronJob;
 /// </summary>
 public sealed record JobExecutionContext
 {
+    internal bool ExecuteChildren = true;
+
     /// <summary>
     /// Represents the context of a job execution. Marked internal to prevent external instantiation.
     /// </summary>
-    /// <param name="jobDefinition">The Job Definition of the Job Run</param>
-    internal JobExecutionContext(JobDefinition jobDefinition)
-    {
-        Parameter = jobDefinition.Parameter;
-        JobDefinition = jobDefinition;
-    }
+    /// <param name="jobRun">The Job Run.</param>
+    internal JobExecutionContext(JobRun jobRun) => JobRun = jobRun;
 
     /// <summary>
     /// Represents the context of a job execution.
     /// </summary>
     public JobExecutionContext(Type jobType, object? parameter)
-        => JobDefinition = new JobDefinition(jobType, parameter, null, null);
+        => JobRun = JobRun.Create(new JobDefinition(jobType, parameter, null, null));
 
     /// <summary>
     /// The Job Instance Identifier, generated once upon creation of the context.
@@ -38,12 +36,31 @@ public sealed record JobExecutionContext
     /// </summary>
     public int Attempts { get; internal set; }
 
-    /// <summary>The Job Definition of the Job Run</summary>
-    internal JobDefinition JobDefinition { get; init; }
+    /// <summary>The Job Run instance.</summary>
+    internal JobRun JobRun { get; }
 
     /// <summary>The Type that represents the Job</summary>
-    internal Type JobType => JobDefinition.Type;
+    internal Type JobType => JobRun.JobDefinition.Type;
 
     /// <summary>The passed in parameters to a job.</summary>
-    public object? Parameter { get; init; }
+    public object? Parameter => JobRun.Parameter;
+
+    /// <summary>
+    /// The correlation identifier of the job run. The <see cref="CorrelationId"/> stays the same for jobs and their dependencies.
+    /// </summary>
+    public Guid CorrelationId => JobRun.CorrelationId;
+
+    /// <summary>
+    /// The output of the parent job. Is always <c>null</c> if the job was not started due to a dependency.
+    /// </summary>
+    public object? ParentOutput => JobRun.ParentOutput;
+
+    /// <summary>
+    /// Prohibits the execution of dependent jobs.
+    /// </summary>
+    /// <remarks>
+    /// Calling <see cref="SkipChildren"/> has no effects when no dependent jobs are defined
+    /// via <see cref="INotificationStage{TJob}.ExecuteWhen"/>.
+    /// </remarks>
+    public void SkipChildren() => ExecuteChildren = false;
 }