+++ src/NCronJob/Registry/JobRegistry.cs
@@ -6,13 +6,13 @@ internal sealed class JobRegistry
 {
     private readonly ImmutableArray<JobDefinition> cronJobs;
     private readonly ImmutableArray<JobDefinition> oneTimeJobs;
-    private readonly ImmutableArray<Type> allJobTypes;
+    private readonly List<JobDefinition> allJobs;
     private readonly ImmutableArray<string> allJobTypeNames;
 
     public JobRegistry(IEnumerable<JobDefinition> jobs)
     {
         var jobDefinitions = jobs as JobDefinition[] ?? jobs.ToArray();
-        allJobTypes = [..jobDefinitions.Select(j => j.Type).Distinct()];
+        allJobs = [..jobDefinitions];
         allJobTypeNames = [.. jobDefinitions.Select(j => j.JobFullName).Distinct()];
         cronJobs = [..jobDefinitions.Where(c => c.CronExpression is not null)];
         oneTimeJobs = [..jobDefinitions.Where(c => c.IsStartupJob)];
@@ -22,7 +22,16 @@ internal sealed class JobRegistry
     public IReadOnlyCollection<JobDefinition> GetAllOneTimeJobs() => oneTimeJobs;
     public IReadOnlyCollection<string> GetAllJobTypes() => allJobTypeNames;
 
-    public bool IsJobRegistered<T>() => allJobTypes.Any(j => j == typeof(T));
+    public bool IsJobRegistered<T>() => allJobs.Exists(j => j.Type == typeof(T));
+
+    public JobDefinition GetJobDefinition<T>() => allJobs.First(j => j.Type == typeof(T));
+
+    public void Add(JobDefinition jobDefinition)
+    {
+        if (!allJobs.Exists(j => j.JobFullName == jobDefinition.JobFullName))
+        {
+            allJobs.Add(jobDefinition);
+        }
+    }
     public int GetJobTypeConcurrencyLimit(string jobTypeName) => cronJobs.FirstOrDefault(j => j.JobFullName == jobTypeName)?.ConcurrencyPolicy?.MaxDegreeOfParallelism ?? 1;
 }
-