+++ src/NCronJob/Scheduler/QueueWorker.LogMessage.cs
@@ -25,4 +25,10 @@ internal sealed partial class QueueWorker
     [LoggerMessage(LogLevel.Trace, "Dequeuing job {JobName} because it has exceeded the expiration period.")]
     private partial void LogDequeuingExpiredJob(string jobName);
 
+    [LoggerMessage(LogLevel.Trace, "Job added to the queue: {JobName} at {ScheduledAt}")]
+    private partial void LogJobAddedToQueue(string jobName, DateTimeOffset? scheduledAt);
+
+    [LoggerMessage(LogLevel.Trace, "Job removed from the queue: {JobName} scheduled for {ScheduledFor}")]
+    private partial void LogJobRemovedFromQueue(string jobName, DateTimeOffset? scheduledFor);
+
 }